Rework u32 to char conversions in the parser
The parser wants to match its u32 sequence against character literals.
Previously it would match `char::from_u32` and then have a bunch of
Some(...) in the match arms. This is quite verbose and it also risks
dropping chars, since the None case should typically be handled the same as
the _ case (that is, a surrogate code point should be handled like any
other code point).

Switch to a new function `to_char_sat` which maps u32s to either their
corresponding char or MAX_CHAR. This allows for handling them more
uniformly.