Fix a bug in case folding characters given by escapes
When we construct a Char node in the ir, we expect the character to be
already folded if our regex is case-insensitive. This was true for
ordinary characters but was not true if the character was given by an
escape.

This happened only in the unoptimizing cases - in the optimizing cases
we unfold the character to determine all possible matches, which masks
the bug. However if regex optimization is disabled then we would panic
in debug builds, and fail to match icase in release builds.