Named captures to return their names in-order
Prior to this commit, the Match::named_groups() function would return
names in an arbitrary order, because they were stored in a hash table.
Switch to returning them in the original specification order from the
pattern. This allows implementing the JS engine semantics where the
object key order matches that from the pattern.

Fixes #96