Revert "Implement caching for unicode unfolding to optimize performance (#119)"
This commit attempted to speed up the optimizer pass for
case-insensitive regexes. However it relies on thread-local storage
which is a big hammer and imposes long-lived memory allocations on
client programmers.

It's not clear how important this caching is; some ad-hoc benchmarks
show unfolds as ~200 nsec which is pretty fast.

If we want to reintroduce caching we should start by creating a
benchmark to justify it; indeed a benchmark for all of regex parsing and
optimization, and then attack what profiling shows is slow.

This reverts commit b333c2fb10fba773bda18d0d58d9452cf5afdf06.